Deploy multiple operand replicas
Currently, we deploy only one replica for both operands
(webhook and controller). This apparently leads to the
operator going unavailable when draining the node where
an operand is running on.

With this patch we run multiple replicas of each operand,
i.e., 1 replica in single-node control-planes or 2 replicas
in multi-node control-planes.

Also, we set maxSurge=0 and maxUnavailable=1 to make sure
nodes are drained one at a time and to prevent more than
one pod running on the same node when nodePorts or
hostNetwork are used.