Add UID to CSO Pod to be able to run with custom SCCs
When there is a custom SCC with "runAsUser: type: runAsNonRoot",
snapshot-operator pod can match it. Then the Pod must have a non-root UID
assigned to be able to run,

This moves the snapshot-operator Pods from "restricted-v2" SCC to
"nonroot-v2" SCC in the default OCP installation.

This is a workaround for the operator using cluster-admin role, which can
allow the operator to use any SCC. In 4.13, we already have proper RBAC
rules for the operator, so this UID is not needed there.