Avoid deleting VolumeGroupSnapshot resources
Prior to this patch, OCP would always try to delete those resources in a
default installation. This is unnecessary because those resources are
only created when TechPreviewNoUpgrade feature set is enabled.

Once this feature set is enabled, it's not possible to go back
to a default installation (i.e., to a state where those resources should
not exist). This means that those resources don't ever
need to be deleted.