Update CRDs from external-snapshotter 8.4.0
That was part of Kubernetes 1.34 release of the controller.

We do not ship the conversion webhook in OCP, therefore disable the
conversion in the CRD. This makes sure OCP installs correctly when TP is
enabled, because the external-snapshotter repo was not rebased yet.

A future PR will disable the v1beta1 APIs